Bug 28622: Update Tor Browser icon for mobile
We follow the desktop layout by providing icons for stable, alpha, and
nightly versions in the official, alpha, and nightly branding
directories, replacing related Firefox icons.

Moreover, this patch adjusts the displayed app name as well according
to the respective Tor Browser flavor being used.

There is no drawable-mdpi directory in Fennec anymore. Thus, we won't
ship new icons for that either and we remove the old torbrowser branding
directory and contents as those are not needed for the new layout.

At some point we could do a proper clean-up of icons we actually don't
need. For instance, right now we still need to ship the
firstrun_welcome.png icon from Fennec's onboarding because otherwise
compilation breaks, even though we have our own, Fennec unrelated
onboarding. But that's part of a different ticket.