Bug 31251: Security Level button UI polish
Changed the anchor to toolbarbutton's anonymous child toolbarbutton-icon
xul:image element so that the security level panel has the same vertical
alignment as other built-in toolbar buttons. Also needed to add the
badged-button class to our toolbar button for this to align properly.

Now setting and deleting the "open" attribute from the toolbarbutton
element so that it maintains the "pushed" style while the security level
panel is open.

Changed the toolbarbutton's oncommand callback to onmousedown to match
the behaviour of firefox's other toolbarbuttons. Security level panel
now opens on mouse press rather than on mouse press and release.

Slightly re-arranged the ordering of the SecurityLevelPanel object's
method definitions so that callbacks appear in a contiguous block to
better match the other objects defined in securityLevel.js

Removed unnecessary async modifier from
SecurityLevelPanel.openAdvancedSecuritySettings().

Fixed a few typos, improved comments, and removed trailing whitespace.